refactored test case formats
