+++ runTests.py
@@ -1,24 +1,31 @@
-import os
 import subprocess
 import sys
 
-testsToRun = [
-    "test.path_constructor.test_path_constructor",
-    "test.logic.test_videoEditing_VideoCropper",
-    "test.logic.test_videoEditing_FirstFrameExtractor",
-    "test.logic.test_videoEditing_merge_and_blend_images",
-    "test.logic.test_videoEditing_merge_and_blend_videos",
-    "test.logic.test_singleton",
-    "test.test_logic.test_customerAttributesEstimator",
-    "test.pages.test_logWindow",
-    "test.pages.test_addStorePopup",
-    "test.pages.test_addStorePopup_ui",
-    "test.pages.test_menuBar_cleanup",
-    "test.pages.test_cameraZoneWindow_videos",
-    "test.pages.test_cameraMergeWindow_ui",
-    "test.pages.test_cameraZoneWindow_ui",
-    "test.integration.test_gui_workflows",
+unitTests = [
+    "test.unit.src.test_path_constructor",
+    "test.unit.src.logic.test_videoEditing_VideoCropper",
+    "test.unit.src.logic.test_videoEditing_FirstFrameExtractor",
+    "test.unit.src.logic.test_videoEditing_merge_and_blend_images",
+    "test.unit.src.logic.test_videoEditing_merge_and_blend_videos",
+    "test.unit.src.logic.test_singleton",
+    "test.unit.src.logic.test_customerAttributesEstimator",
+    "test.unit.src.pages.test_logWindow",
+    "test.unit.src.pages.test_addStorePopup",
+    "test.unit.src.pages.test_menuBar_cleanup",
+    "test.unit.src.pages.test_cameraZoneWindow_videos",
 ]
+
+customerTests = [
+    "test.customer.src.pages.test_addStorePopup_ui",
+    "test.customer.src.pages.test_cameraMergeWindow_ui",
+    "test.customer.src.pages.test_cameraZoneWindow_ui",
+]
+
+integrationTests = [
+    "test.integration.src.pages.test_gui_workflows",
+]
+
+testsToRun = unitTests + customerTests + integrationTests
 
 
 def main():
