+++ test/customer/src/pages/gui_test_utils.py
