+++ test/customer/src/pages/test_addStorePopup_ui.py
@@ -5,7 +5,7 @@ import dearpygui.dearpygui as dpg
 
 from src.database import model_managers as mm
 from src.pages import addStorePopup
-from test.pages.gui_test_utils import GuiDbTestCase
+from test.customer.src.pages.gui_test_utils import GuiDbTestCase
 
 
 class TestAddStorePopupUI(GuiDbTestCase):
