+++ test/customer/src/pages/test_cameraMergeWindow_ui.py
@@ -7,7 +7,7 @@ import dearpygui.dearpygui as dpg
 from src.database import model_managers as mm
 from src.database.models import Store
 from src.pages import cameraMergeWindow
-from test.pages.gui_test_utils import GuiDbTestCase
+from test.customer.src.pages.gui_test_utils import GuiDbTestCase
 
 
 class TestCameraMergeWindowUI(GuiDbTestCase):
