+++ test/integration/src/pages/test_gui_workflows.py
@@ -8,7 +8,7 @@ import dearpygui.dearpygui as dpg
 from src.database import model_managers as mm
 from src.database.models import Camera, Store
 from src.pages import addStorePopup, cameraMergeWindow, cameraZoneWindow, menuBar
-from test.pages.gui_test_utils import GuiDbTestCase
+from test.customer.src.pages.gui_test_utils import GuiDbTestCase
 
 
 class TestGuiWorkflows(GuiDbTestCase):
