+++ test/unit/src/logic/test_singleton.py
