+++ test/unit/src/logic/test_videoEditing_FirstFrameExtractor.py
