+++ test/unit/src/logic/test_videoEditing_VideoCropper.py
