+++ test/unit/src/logic/test_videoEditing_merge_and_blend_images.py
