+++ test/unit/src/pages/test_addStorePopup.py
