+++ test/unit/src/pages/test_logWindow.py
